SOF-303: Consolidate specimen image overlay UI into separate component
